Fix end loop templates starting too late

+++ src/core/DevilDaggersInfo.Core.Spawnset/View/SpawnView.cs
@@ -1,3 +1,3 @@
 namespace DevilDaggersInfo.Core.Spawnset.View;
 
-public readonly record struct SpawnView(EnemyType EnemyType, double Seconds, int NoFarmGems, GemState GemState);
+public readonly record struct SpawnView(EnemyType EnemyType, float Seconds, int NoFarmGems, GemState GemState);
